+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This project uses [Semantic Versioning 2.0.0](http://semver.org/), the format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+
+- Add zoom indicator showing current zoom percentage when zoomed or panned.
+- Add zoom in/out buttons to the view action bar for precise zoom control.
+
 ## 0.3.2 - 2026-01-29
 
 ### Changed